+++ packages/assignment-7/src/App.tsx
@@ -55,7 +55,7 @@ interface RepeatInfo {
   endDate?: string;
 }
 
-interface Event {
+export interface Event {
   id: number;
   title: string;
   date: string;
